+++ RedPandaIDE/test/resources/preprocessor-macros-2-result.cpp
@@ -1,0 +1,6 @@
+#define FUNCTION(name,a) int fun_##name() { return a; }
+
+int fun_abcd() { return 12; }
+
+
+int x;