+++ RedPandaIDE/test/resources/preprocessor-macros-2.cpp
@@ -1,0 +1,6 @@
+#define FUNCTION(name,a) int fun_##name() { return a; }
+
+FUNCTION(
+abcd, 12
+)
+int x;