+++ RedPandaIDE/test/test_cpppreprocessor.cpp
@@ -83,6 +83,30 @@ void TestCppPreprocessor::test_macro_replace_5()
              preprocessor.expandMacros("ttt TEST TEST"));
 }
 
+void TestCppPreprocessor::test_macro_replace_6()
+{
+    CppPreprocessor preprocessor;
+    QFileInfo info("resources/preprocessor-macros-2.cpp");
+    preprocessor.preprocess(info.absoluteFilePath());
+    QStringList text1 = filterIncludes(preprocessor.result());
+    QStringList text2 = readFileToLines("resources/preprocessor-macros-2-result.cpp");
+    QVERIFY(!text1.isEmpty());
+    QVERIFY(!text2.isEmpty());
+    QCOMPARE(text1,text2);
+}
+
+void TestCppPreprocessor::test_macro_replace_7()
+{
+    CppPreprocessor preprocessor;
+    QFileInfo info("resources/preprocessor-macros-3.cpp");
+    preprocessor.preprocess(info.absoluteFilePath());
+    QStringList text1 = filterIncludes(preprocessor.result());
+    QStringList text2 = readFileToLines("resources/preprocessor-macros-3-result.cpp");
+    QVERIFY(!text1.isEmpty());
+    QVERIFY(!text2.isEmpty());
+    QCOMPARE(text1,text2);
+}
+
 QStringList TestCppPreprocessor::filterIncludes(const QStringList &text)
 {
     QStringList result;